implement abstract database layer with janusgraph just one of them
